feat: refactor SelectLine component and styles, integrate new FormSelectLine, and enhance MessageItem and MessageList styling
