+++ app/components/ChatBoot/ChatBoot.module.scss
@@ -1,9 +1,6 @@
 .chatBootContainer {
   display: flex;
   flex-direction: column;
-  min-height: 100vh;
-  height: 100vh;
   background: none;
-  overflow-y: auto;
   padding-bottom: 0.5rem;
 }
