+++ app/components/Header/Header.module.scss
@@ -2,7 +2,6 @@
   width: 100%;
   background: #fff;
   box-shadow: 0 2px 8px rgba(83, 181, 230, 0.08);
-  border-bottom: 1px solid #53b5e6;
   border-radius: 12px;
 }
 
@@ -12,7 +11,7 @@
   justify-content: space-around;
   align-items: center;
   padding: 1rem;
-  border: 1px solid #53b5e6;
+
 }
 
 .logo {
