+++ app/components/MessageItem/MessageItem.module.scss
@@ -1,0 +1,46 @@
+.messageItem {
+  margin-bottom: 16px;
+}
+
+.imagesWrapper {
+  max-width: 600px;
+  margin: 0 auto;
+  padding: 0 40px;
+}
+
+.citationsWrapper {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 12px;
+  margin: 16px 0;
+}
+
+.citationLink {
+  display: inline-block;
+  padding: 10px 18px;
+  background: #eaf6ff;
+  border-radius: 8px;
+  color: #0074cc;
+  font-weight: 500;
+  text-decoration: none;
+  box-shadow: 0 1px 4px rgba(0, 0, 0, 0.04);
+  transition: background 0.2s;
+  margin-bottom: 4px;
+}
+
+.markdownText {
+  color: #fff;
+}
+
+.tableTh {
+  border: 1px solid #ccc;
+  padding: 8px;
+  background: #f3f3f3;
+  font-weight: bold;
+}
+
+.tableTd {
+  border: 1px solid #ccc;
+  padding: 8px;
+  color: #111;
+}
