+++ app/components/MessageList/MessageList.module.scss
@@ -1,0 +1,18 @@
+.messageListContainer {
+  margin-top: 1rem;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  padding: 16px;
+  min-height: 300px;
+  margin-bottom: 12px;
+
+  background: none;
+}
+
+.emptyMessage {
+  margin-top: 24px;
+  margin-bottom: 24px;
+  text-align: center;
+  font-size: 1.1rem;
+  color: #fff;
+}
